refactor(bixarena): extract helix geometry, drop category filter, add responsive prompt-card CTA
