+++ libs/bixarena/ui/src/lib/prompt-card/prompt-card.component.html
@@ -11,7 +11,7 @@
         <span class="prompt-card-stat">{{ count() }} {{ countLabel() }}</span>
       }
       <span class="prompt-card-cta">
-        {{ ctaText() }}
+        <span class="prompt-card-cta-text">{{ ctaText() }}</span>
         <svg
           class="prompt-card-cta-icon"
           viewBox="0 0 24 24"
