+++ libs/bixarena/ui/src/lib/prompt-card/prompt-card.component.scss
@@ -1,3 +1,5 @@
+@use 'shared/typescript/styles/src/shared-styles/variables' as shared;
+
 :host {
   display: block;
 }
@@ -14,6 +16,7 @@
 .prompt-card-meta {
   display: flex;
   align-items: center;
+  flex-wrap: wrap;
   gap: 0.75rem;
 }
 
@@ -25,6 +28,7 @@
   padding: 0.125rem 0.5rem;
   border-radius: 999px;
   margin-left: -0.25rem;
+  white-space: nowrap;
 }
 
 .prompt-card-stat {
@@ -41,11 +45,18 @@
   color: var(--p-text-muted-color);
   transition: color 0.15s;
   margin-left: auto;
+  white-space: nowrap;
 }
 
 .prompt-card-cta-icon {
   width: 0.875rem;
   height: 0.875rem;
+}
+
+@media (#{shared.$lg-breakpoint} < width <= #{shared.$xl-breakpoint}) {
+  .prompt-card-cta-text {
+    display: none;
+  }
 }
 
 .prompt-card {
